Generate entity name types with a macro rule
Entities on GitHub can have a human-readable name that identifies them.
The names are represented by their own types in github-parts, which
created a lot of code duplication. By generating the types from a macro
rule, a lot of code can be removed from the project.
