+++ src/lib.rs
@@ -8,36 +8,7 @@
 
 #![warn(missing_docs)]
 
-macro_rules! id {
-    ($name:ident) => {
-        /// Id
-        ///
-        /// Resources on GitHub have a unique `id` that is used to interact with them through
-        /// GitHub's REST API.
-        #[derive(
-            Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash, Debug, Deserialize, Serialize,
-        )]
-        pub struct $name(u64);
-
-        impl $name {
-            /// Initializes a new id.
-            pub fn new(id: u64) -> Self {
-                Self(id)
-            }
-
-            /// Returns the id.
-            pub fn get(&self) -> u64 {
-                self.0
-            }
-        }
-
-        impl Display for $name {
-            fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-                write!(f, "{}", self.0)
-            }
-        }
-    };
-}
+mod macros;
 
 pub mod account;
 pub mod action;
